Add symbolically perturbed exact circle csg
Circle csg turns out to involve quite large predicates: the current
code uses degree 20, although this is reducible to degree 12 with
no algorithmic changes and degree 8 if ray firing is adjusted to use
rays at integer heights.  In addition, representing nearly straight
lines with circular arcs sacrifices nearly half the quantization bits
in order to place very large radius circles very far away.

For both of these reasons, the ExactInt type has been bumped from
int32_t to int64_t, with quantized values in the range (-2**53,2**25).
This causes a potentially significant slowdown, but correctness wins.

+++ exact/test_perturb.py
@@ -16,7 +16,7 @@ def test_interpolation():
   for d in xrange(1,5):
     for n in xrange(1,6):
       beta = perturb_monomials(d,n)
-      coefs = random.randint(2*bound,size=len(beta)).astype(int32)-bound
+      coefs = random.randint(2*bound,size=len(beta))-bound
       in_place_interpolating_polynomial_test(d,beta,coefs,False)
 
 def test_perturbed_sign():
